cleaned up the daq to work in a smart way - controller manages the actual recording while the window just calls controller functions and flips its button states
